added support for per-component sosflow setup.
'codes' takes another key 'sosflow' that can be set to True/False to indicate if a code has been built with tau/sosflow.
If a code is not built with sosflow, its nodes will not be counted during calculation of sosflow listener offsets.

The sosflow=True/False bool option for sweep groups is still maintained to enable/disable sosflow for a group