Added support for sos analysis (generating vtk files from sos data) and support for multiple sos aggregators.
Added new sweepgroup parameter 'sos_analysis' that denotes if the sosflow analysis code must be run to generate vtk data.
Added 'sos_analysis_path' as a variable to a campaign defn.
Function 'insert_sosflow' will add the analysis code to the list of codes to be run per Run object.