Fixed bug where the adios xml transport and transform functions in Cheetah were writing back the application's xml file without first line containing the xml declaration.
Without the xml declaration, the xml parser in adios was ignoring the paramters being passed to the MPI_AGGREGATE transport method and was setting all ranks as aggregators as default.
This led to sub-optimal I/O performance.

As a fix, now passing the xml_declaration argument to the ElementTree.write() function. Leaving the default encoding to US-ASCII.