use TERM signal instead of KILL
fixes #83

This should give the aprun wrapper a chance to send the signal
on to real aprun so it can send to actual application code and
exit cleanly. If the application code is truly hung, this may not
work, but using a process group with Popen start_new_session does
not work well with the theta aprun wrapper.