Resolves #138: ability to chain RCs using dependencies.
Introduced a new property for the SweepGroup class - 'rc_dependency',
which is a dict where the key represents the RC that depends on the list
of RCs in the dict value.

Currently supporting simple dependencies where an RC waits for another
RC to finish, as opposed to more complex dependency mgmt such as
after_ok, after_done etc.